requirements.txt
pandas
PyMuPDF
faiss-cpu
sentence-transformers
flask
flask-cors
ollama
firebase-admin
fastapi
uvicorn[standard]
python-multipart
aiofiles
Flask
Flask-Cors
pymongo
Werkzeug
PyJWT
Flask-Session
pymongo
python-dotenv
werkzeug
flask-cors